requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyMuPDF>=1.24.11
python-dotenv==1.0.0
google-genai
requests